+++ CHANGELOG.md
@@ -1,7 +1,18 @@
 ## v1.14.0 (unreleased)
+FEATURES:
+- **New Data Source**: azapi_client_config
+
+ENHANCEMENTS:
+- `azapi` provider: Support `client_certificate` field, which specifies base64-encoded PKCS#12 bundle to be used as the client certificate for authentication.
+- `azapi` provider: Support `use_aks_workload_identity` field to enable AKS workload identity.
+- `azapi_resource`, `azapi_update_resource`, `azapi_resource_action`, `azapi_data_plane_resource` resources: Support `timeouts.update` field, which is used to specify the timeout for the update operation.
+- `azapi_update_resource` resource: Improve the id build logic to honor user's input.
 
 BUG FIXES:
 - Fix a bug that `azapi_resource` will crash when the `location` in GET response is null.
+- Fix a bug that schema validation fails to validate unknown string values.
+- Fix a bug that `azapi_resource` only supports extension resource on a resource group scoped resource.
+- Fix a bug that OIDC authentication should not be used when required fields are missing.
 
 ## v1.13.1